+++ packages/core/src/types/router.interface.ts
@@ -2,6 +2,8 @@ import type { IgniterControllerConfig } from "./controller.interface";
 import type { DocsConfig, IgniterBaseConfig } from "./builder.interface";
 import type { ContextCallback } from "./context.interface";
 import type { MutationActionCallerResult, QueryActionCallerResult } from "./client.interface";
+import type { IgniterAction } from "./action.interface";
+import type { Prettify } from "./utils.interface";
 
 export type IgniterRouterCaller<
   TControllers extends Record<string, IgniterControllerConfig<any>>, // ✅ Simplificado
@@ -10,7 +12,6 @@ export type IgniterRouterCaller<
     [A in keyof TControllers[C]['actions']]:
     TControllers[C]['actions'][A]['type'] extends 'query' ? {
       type: 'query';
-      useQuery: (...args: any[]) => QueryActionCallerResult<TControllers[C]['actions'][A]>
       query: (input: any) => Promise<TControllers[C]['actions'][A]['$Infer']['$Response']>
     } : {
       type: 'mutation';
@@ -19,6 +20,40 @@ export type IgniterRouterCaller<
     }
   }
 }
+
+export type ServerExtraCallerInput = {
+  headers?: Record<string, string> | undefined;
+  cookies?: Record<string, string>;
+  credentials?: RequestCredentials;
+}
+
+export type InferServerRouterCallerAction<
+  TAction extends IgniterAction<any, any, any, any, any, any, any, any, any, any>,
+  TCaller = TAction['$Infer']['$Caller'],
+  TCallerParams = TCaller extends (input: infer P) => any ? P : never,
+  TCallerReturn = TCaller extends (input: any) => Promise<infer R> ? R : never
+> = TAction extends { method: "GET" }
+  ? {
+      type: 'query';
+      query: (input: Prettify<TCallerParams & ServerExtraCallerInput>) => Promise<TCallerReturn>;
+    }
+  : {
+      type: 'mutation';
+      mutate: (input: Prettify<TCallerParams & ServerExtraCallerInput>) => Promise<TCallerReturn>;
+    };
+
+export type InferServerRouterCaller<
+  TRouter extends IgniterRouter<any, any, any, any, any>,
+> =
+  TRouter extends IgniterRouter<any, infer TControllers, any, any, any>
+    ? {
+        [TControllerName in keyof TControllers]: {
+          [TActionName in keyof TControllers[TControllerName]["actions"]]: InferServerRouterCallerAction<
+            TControllers[TControllerName]["actions"][TActionName]
+          >;
+        };
+      }
+    : never;
 
 export type IgniterRouterConfig<
   TContext extends object | ContextCallback,
@@ -39,12 +74,14 @@ export type IgniterRouter<
   TControllers extends Record<string, IgniterControllerConfig<any>>, // ✅ Simplificado
   TConfig extends IgniterBaseConfig,
   TPlugins extends Record<string, any>,
-  TDocs extends DocsConfig
+  TDocs extends DocsConfig,
 > = {
   config: TConfig & { docs?: TDocs };
   controllers: TControllers;
   handler: (request: Request) => Promise<Response>;
-  $context: TContext;
-  $plugins: TPlugins;
-  caller: IgniterRouterCaller<TControllers>;
+  caller: InferServerRouterCaller<IgniterRouter<TContext, TControllers, TConfig, TPlugins, TDocs>>;
+  $Infer: {
+    $context: TContext;
+    $plugins: TPlugins;
+  }
 }